dbComplited

+++ PCKonfiguratorBackend/Models/CPU/CPU.cs
@@ -8,7 +8,7 @@ namespace PCKonfiguratorBackend.Models
         public Guid id { get; set; }
         public string name { get; set; }
         public string manufacturer { get; set; }
-        public decimal price { get; set; }
+        public float price { get; set; }
         public string image { get; set; }
         public CPUSpecification cpuSpecification {  get; set; }
     }
