+++ PCKonfiguratorBackend/Models/GPU/GPUSpecifications.cs
@@ -8,7 +8,7 @@ namespace PCKonfiguratorBackend.Models
         [Key]
         public Guid id { get; set; }
         public int memory { get; set; }
-        public decimal speed { get; set; }
+        public float speed { get; set; }
         public int boostclock { get; set; }
         public int coreclock { get; set; }
         public int power {  get; set; }
